refact: useSearch custom hook사용하여 책임 분리
